Refactor MainActivity to use MainScreen composable
The MainActivity was refactored to use a new MainScreen composable.
The MainScreen now displays a list of Chinese words fetched from the MainScreenViewModel.
It includes a filter for HSK levels.
The WordItem composable displays details of each word.
Added Hilt dependency to the project.
Removed unused code from MainActivity.kt